+++ src/Localization/_LOCALIZATION_MODULE.mjs
@@ -1,1 +1,0 @@
-export const PWA_LOCALIZATION_MODULE = "pwa";