+++ src/SettingsStorage/SettingsStorage.mjs
@@ -1,0 +1,3 @@
+/**
+ * @typedef {{get: (key: string, default_value: *) => Promise<*>, store: (key: string, value: *) => Promise<void>}} SettingsStorage
+ */